Use docker ps to get containers
Use docker ps with filter parameter to get the containers related to a
docker-compose scenario. This allows us to avoid loading environment
variables and the profiles to get this information.
